Add files via upload
Login pages and the stock have the basic functionality, but I can still advanced these pages if necessary and perhaps adding to each of the categories for the stock page to make the user interface interactive, so the customer can keep track of the orders they made. I also have included form validation functionality to ensure that the customer cannot login with empty fields that are required. Elements that I styled (using CSS) are embedded into HTML and didn't create a CSS file for maintainability. I found it easier to embed CSS into HTML because I can keep track of individual elements styled accordingly a lot better.